Update patch set 3
Patch Set 3: Code-Review-1

(2 comments)

I'm really not clear on what the exact distinction between GEOPM_REGION_ID_EPOCH, GEOPM_REGION_ID_UNMARKED and GEOPM_REGION_ID_UNDEFINED.

My guess is that the EPOCH does not have a real Region object, but is simply used to transfer beacons from the application across shared memory.  UNMARKED is code that is outside of a marked region, so I would think that as long as there is one sample, we have entered the UNMARKED region.  UNDEFINED is used as a designation for an unset region is a memory to mark it as unknown.

Patch-set: 3
Label: Code-Review=-1