Update patch set 3
Abandoned

Chris convinced me this was a bad idea.  The hup handler will cancel all jobs that match on a combination of the user id and job name.  A user might have another big job running that this would kill.

This is only a problem when running from the command line, so we really just always be running this through a batch script.

Patch-set: 3
Status: abandoned
Tag: autogenerated:gerrit:abandon