Adding initial support to account for HFI power on each node. In this commit, the HFI power for a given node is subtracted from its total power budget as indicated by the parent controller.
This commit assumes that there is a linear relationship
between the bandwidth and the power consumed by the
HFI. This approximation is tolerable because HFI power
is more than an order of magnitude lesser than node
power. The two extremities of the power v/s bandwidth
curve correspond to the Post-Si estimates of WFR
(KNL/SKX) for idle and typical workloads.

Change-Id: I1a139541e094e1fa89ffb1102b382935c6b0d3e1
Signed-off-by: Siddhartha Jana <siddhartha.jana@intel.com>