Update patch set 2
Patch Set 2:

(4 comments)

Our software as a whole is targeting bulk synchronous applications and the app behavior you've described doesn't quite fit that category/model.  That being said there are gotcha/corner cases.  The contract that we intend to implement is one which the average calculation only takes into account the most recent rank transitions in a region with hopes of preventing old transitions being recorded in the average and generating noise for deciders making decisions on these values.

Patch-set: 2