Update patch set 1
Patch Set 1:

> Patch Set 1: Code-Review-1
> 
> Changes need to be made in other places in the accounting (in particular in Profile::exit()).  Please look into this solution for HACC.

So I've been staring at this for a while, and don't see necessary changes in Profile::exit() (or anywhere else for that matter).  I've managed to convince myself this will work as-is.

I'm going to do some testing with it to see how it goes.

Patch-set: 1
Reviewer: Gerrit User 1009284 <1009284@d5d70762-12d0-45a1-890d-524b12d3f735>