Update patch set 3
Patch Set 3: Code-Review+1

The change looks good. I think seeing a class name as argv[1] is unexpected for a launch command, but I see the reason you're doing it is to map the dummy launcher to exactly whatever the existing launcher is. Maybe there's a different way the API could be organized to avoid this kind of confusion in the future.

Patch-set: 3
Reviewer: Gerrit User 1016629 <1016629@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Code-Review=+1