Update patch set 3
Patch Set 3:

If the duplicated code you're referring to is just about get_alloc_nodes() not being supported in the new launcher classes, please leave that code duplicated.

I think it makes more sense to have it unimplemented in the base class.  It's expected that if get_alloc_nodes() is not supported in one of the derived classes that the code will be duplicated.

Patch-set: 3
Reviewer: Gerrit User 1009284 <1009284@d5d70762-12d0-45a1-890d-524b12d3f735>