Update patch set 6
Patch Set 6:

> Patch Set 6:
> 
> > Patch Set 6: Code-Review-1
> > 
> > (1 comment)
> 
> I don't understand how you can say that when it just resides with the rest of the run config logic... Just following the existing pattern for 1 and secondly the node list is loosely dependent on the number of nodes requested for the job (does truncation of the list if your allocation is larger than the current node request).

The rest of the run config logic passes config parameters that the launcher actually needs in order to run.  In every case where set_node_list() is called, you're relying on the default behavior of the method which has node_list set to None and then only takes action if the launcher is ompi.  In every other case this is a no-op and self._node_list is None.

This is messy the way that it is.  I understand the point about node list being dependent on the number of nodes requested.  Since that isn't something passed in at construction time, the call to set_node_list() cannot go there.  I think it would be safe to move it to the beginning of test launcher's run() method.  The tests themselves do not need to be aware of this requirement to set the node list.  The launcher can handle that for them.

Patch-set: 6
Reviewer: Gerrit User 1009284 <1009284@d5d70762-12d0-45a1-890d-524b12d3f735>