Update patch set 8
Patch Set 8:

Now that I'm writing the documentation for this class, I feel more and more that this logic is not enough to deserve an entire class.  The only advantage is that we can write unit tests of this one function, which is the bulk of the slurm plugin.  The same behavior should also be covered by the integration test (next patch).

Patch-set: 8
Reviewer: Gerrit User 1012567 <1012567@d5d70762-12d0-45a1-890d-524b12d3f735>