Update patch set 2
Patch Set 2: Code-Review-1

Need continued debug of the new test app, something seems fishy wrt the thread table.  In the event the app and runtime execute completely and a report comes out the other end, the observed region count is 1 not the expected 50.  The issue seems to be exacerbated when an MPI collective is added to the mix (next patch in stack).  Is this a valid app configuration?  If so it seems like we have a bug we need to triage and fix.

Patch-set: 2
Reviewer: Gerrit User 1011138 <1011138@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Code-Review=-1