Improve test coverage for `ServiceIOGroupTest`:
 - `ServiceIOGroupTest.read_signal_exception`
 - `ServiceIOGroupTest.write_control_exception`

Remove the duplicate test fixtures of `ServiceIOGroupTest`

 - `ServiceIOGroupTest.save_control`
 - `ServiceIOGroupTest.restore_control`

The point of these tests is to show that the method does not do anything.
If we send it an invalid path, we know that if it tries to "do something" i.e. open that file, the test will fail.
The `"/bad/file/path"` is prefered since we can also assume that it cannot be opened.

Also remove some test fixtures that we are not having any more from `test/Makefile.mk`

 - `ServiceIOGroupTest.sample`
 - `ServiceIOGroupTest.adjust`

Signed-off-by: Konstantin Rebrov <konstantin.rebrov@intel.com>