Adding `BatchServerTest.fork_and_terminate_child` and `BatchServerTest.fork_and_terminate_parent`
 - `BatchServerTest::fork_other()` to support switching the server and client between the child process and parent process at will.

These are more than just unit tests in the sense that a fork happens, and there is a child process and a parent process.
One of these processes becomes the server, the other one the client.
These tests check the interaction between the server and the client, when the client sends the server a `BatchStatus::M_MESSAGE_TERMINATE` or `SIGTERM`. So they are more accurate use case than `BatchServerTest.receive_message_terminate`

 - `BatchServerTest.fork_and_terminate_child` checks the situation when the server is on the child process.
   - Simulates the usual way that the `BatchServer` gets a terminate message from the user.
   - NOTE: This test only improves testing of use case. It doesn't improve the coverage. This is because LCOV is not tracking the child process.
 - `BatchServerTest.fork_and_terminate_parent` checks the situation when the server is on the parent process.
   - This has the added benefit that LCOV is able to provide coverage for the lines in the server, such as registering the signal handler, and calling it when a `SIGTERM` is received from the client.

Here we are also using a custom mock for `m_batch_status->receive_message()` takes a lambda function as a behavior which simulates blocking on the `read()` system call, and then throws an `geopm::Exception` when the `SIGTERM` is received.

Misc edits to `BatchServerTest.cpp`

 - add `EXPECT_CALL()` for `m_batch_status->send_message()` in `BatchServerTest.read_batch_exception`

Signed-off-by: Konstantin Rebrov <konstantin.rebrov@intel.com>