Resolve leaked mock objects using `_Exit()` instead of `testing::Mock::AllowLeak()` and `this->TearDown()`
The underlying problem was that **Google Test** was reporting leaked mock objects in the forked process (child process). To mitigate this we previously used `testing::Mock::AllowLeak()` and `this->TearDown()`. Now instead we are using `_Exit()` in the forked process, which stops reporting leaked mock objects.
The `_Exit()` function in C/C++ gives normal termination of a program without performing any cleanup tasks.

Signed-off-by: Konstantin Rebrov <konstantin.rebrov@intel.com>