Modify `conf.py` to generate html/man pages from the rst files, without the ronn files
Previously the `conf.py` used to generate rst files from the ronn files,
and then generate the html and roff (man) files from the rst files.

Now since the ronn files have been deleted, there is no longer any dependency on them in the `conf.py`.
Additionally this prevents the rst being re-generated, hence overwriting necessary the manual edits to them.

This commit lists the man pages to be generated in the `rst_files`
variable directly instead of getting that list from the directory.

Also this commit removes all the python functions that used scripting to
modify the rst file immediately after it was generated, and related ones.

Instead, this commit introduces `get_description()` function, which gets
the **description text** from the rst file itself, and puts it into the
roff file. Previously there was a bug that the man pages did not include
the right description text. This commit fixes the bug.

Signed-off-by: Konstantin Rebrov <konstantin.rebrov@intel.com>