Correct ActiveSessions implementation to reflect new tests
Refactor the `__init__()` of `ActiveSessions` in varrun.py

Make the conditions which determine how and when to create a directory
to hold the session files, more robust.

The directory is created if it doesn't already exist, or if it exists in
an invalid state, which could be one of the following:
 - It's a link not a directory.
 - It's any other kind of file not a directory.
 - The directory's permissions is not what we wanted.
 - The directory's user owner is not what we wanted.
 - The directory's group owner is not what we wanted.

In that case it is renamed, and a new directory is created in it's place
with the desired name.

This commit also contains modifications to `_load_session_file()` to
facilitate printing extra verbose warning messages.

Added new tests to TestActiveSessions.py

 - test_creation_link_not_dir()
 - test_creation_file_not_dir()
 - test_creation_bad_perms()
 - test_creation_bad_user_owner()
 - test_creation_bad_group_owner()
 - test_creation_sessions()
 - test_creation_bad_session_perms()
 - test_creation_bad_session_user_owner()
 - test_creation_bad_session_group_owner()

Signed-off-by: Konstantin Rebrov <konstantin.rebrov@intel.com>