Implement `secure_read_file()` and remove `secure_remove_file()`
Write implementation for `secure_read_file()` in `varrun.py` and replace
the original code in `_load_session_file()` with a call to the function.

Delete the `secure_remove_file()`, which we no longer need.

Modify the following tests in the `TestActiveSessions.py` to make them
pass with the new implementation of `secure_read_file()`.

- test_creation_bad_session_perms
- test_creation_bad_session_user_owner
- test_creation_bad_session_group_owner

Signed-off-by: Konstantin Rebrov <konstantin.rebrov@intel.com>