Add checks to the other functions used in `mock.patch`
Previously only `sys.stderr.write` had a named object in `mock.patch`
and the calls to that function were checked.
Now we are having named objects and checked calls for all the other
functions that are used in `mock.patch`

Also updated some of the tests to use more asserts for robustness.

Signed-off-by: Konstantin Rebrov <konstantin.rebrov@intel.com>