Prevent `secure_read_file()` from opening a fifo
Previously the function `secure_read_file()` used to blindly open a fifo,
if it was provided one as the `path`, causing the process to block until
something came through the other end of the fifo. This **bug** was
causing the program to hang indefinitely.

Now I added a check so that `secure_read_file()` prints a warning
message if it determines that the passed in `path` is a fifo, and it
doesn't open it.

This change enables test_read_file_fifo in TestSecureFiles to use an
actual fifo as the input, instead of a regular file which had been
cleverly mocked to make it look like a fifo. The test is now upadted to
reflect this change.

Signed-off-by: Konstantin Rebrov <konstantin.rebrov@intel.com>